requirements.txt
alembic==1.13.0
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
Faker==19.8.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.0.2
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
pypdf==3.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.49
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==2.3.7
WTForms==3.0.1
XLML==0.1.2
